enable raycast to object

+++ include/navi_sim/primitives/box.hpp
@@ -22,7 +22,7 @@ namespace navi_sim
 class Box : public Primitive
 {
 public:
-  explicit Box(double depth, double width, double height, geometry_msgs::msg::Pose pose);
+  explicit Box(float depth, float width, float height, geometry_msgs::msg::Pose pose);
 
 private:
 };
